+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* `lowlight` when the mouse leaves the configurator area
 
 ## [1.25.7] - 2021-04-29